Fix the dependency of accounts_file.rs (#30423)
Replace the use of append_vec::* in accounts_file.rs
